requirements.txt
colorama==0.4.3
Pillow==7.1.2
